requirements.txt
fastapi
uvicorn
python-multipart
mysql-connector-python
pydantic
airefinery-sdk
pyx12
pymongo
google-cloud-storage
xmltodict
